KG-RAG-master/requirements.txt
bm25s==0.2.8
langchain==0.3.23
langchain_chroma==0.2.3
langchain_community==0.3.21
langchain_core==0.3.54
langchain_ollama==0.3.2
langchain_text_splitters==0.3.8
matplotlib==3.10.1
numpy==2.2.4
PyStemmer==2.2.0.3
PyYAML==6.0.2
scikit_learn==1.6.1
setuptools==75.8.0